GFS2: move toward a generic multi-block allocator
Patchwork-id: 44936
O-Subject: [RHEL6.3 PATCH 1/7] GFS2: move toward a generic multi-block allocator
Bugzilla: 788627
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

From: Bob Peterson <rpeterso@redhat.com>

This patch is a revision of the one I previously posted.
I tried to integrate all the suggestions Steve gave.
The purpose of the patch is to change function gfs2_alloc_block
(allocate either a dinode block or an extent of data blocks)
to a more generic gfs2_alloc_blocks function that can
allocate both a dinode _and_ an extent of data blocks in the
same call. This will ultimately help us create a multi-block
reservation scheme to reduce file fragmentation.

This patch moves more toward a generic multi-block allocator that
takes a pointer to the number of data blocks to allocate, plus whether
or not to allocate a dinode. In theory, it could be called to allocate
(1) a single dinode block, (2) a group of one or more data blocks, or
(3) a dinode plus several data blocks.

rhbz#788627

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>