audit: allow interfield comparison in audit rules
Patchwork-id: 44955
O-Subject: [PATCH RHEL6.3 05/10] audit: allow interfield comparison in audit
	rules
Bugzilla: 661765
RH-Acked-by: Dave Anderson <anderson@redhat.com>

RHBZ: 661765
Upstream commit: 02d86a5

We wish to be able to audit when a uid=500 task accesses a file which is
uid=0.  Or vice versa.  This patch introduces a new audit filter type
AUDIT_FIELD_COMPARE which takes as an 'enum' which indicates which fields
should be compared.  At this point we only define the task->uid vs
inode->uid, but other comparisons can be added.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>