Fix sysctl check failure update
Patchwork-id: 45061
O-Subject: [RHEL6.3 v2 PATCH] Fix sysctl check failure
Bugzilla: 714021
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=714021

This is a revised version of the patch previously
posted for this bug.

This bug was seen while booting kernel-debug in a kvm guest.  There were
some cherry-picked changes to fix Bug 688725 that led to this.  The actual
RHEL6 commit is cdbb8035b81a11922610966da3ffbd0f0c66cf30 .  It turns
out that that particular commit was not necessary to the operation of
the driver; it was part of a revision of the sysctl facility to eliminate
the legacy binary interface.  We're not about to do this in RHEL6 because
it breaks KABI.

>From the BZ comments:

"When md_init calls register_sysctl_table with this raid_root_table,
eventually it gets checked by sysctl_check_dir.  This will look up
the namespace in the root_table and compare the procname and ctl_name
between it and the table being registerd, and if they don't match,
it causes the failure seen in this bz."

Complete details are in Comment 7 of the bug

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>