vmxnet3: Cap the length of the pskb_may_pull on transmit (bz 790673)
Patchwork-id: 45091
O-Subject: [RHEL 6.3 PATCH] vmxnet3: Cap the length of the pskb_may_pull on
	transmit (bz 790673)
Bugzilla: 790673
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

From: Neil Horman <nhorman@shamino.rdu.redhat.com>

Backport of a pending upstream commit:

I was recently shown that vmxnet3 devices on transmit, will drop very small udp
frames consistently.  This is due to a regression introduced by commit
39d4a96fd7d2926e46151adbd18b810aeeea8ec0.  This commit attempts to introduce an
optimization to the tx path, indicating that the underlying hardware behaves
optimally when at least 54 bytes of header data are available for direct access.
This causes problems however, if the entire frame is less than 54 bytes long.
The subsequent pskb_may_pull in vmxnet3_parse_and_copy_hdr fails, causing an
error return code, which leads to vmxnet3_tq_xmit dropping the frame.

Fix it by placing a cap on the copy length.  For frames longer than 54 bytes, we
do the pull as we normally would.  If the frame is shorter than that, copy the
whole frame, but no more.  This ensures that we still get the optimization for
qualifying frames, but don't do any damange for frames that are too short.

Testing:
GSS and reporting customer have confirmed that this patch solves the reported
problem

Upstream:
Sent, waiting on review:
http://marc.info/?l=linux-netdev&m=132939301126255&w=2

Brew:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4051348
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>