mlx4: Add support for promiscuous mode in the new steering model.
Patchwork-id: 45096
O-Subject: [Patch RHEL6 003/216] mlx4: Add support for promiscuous mode in the
	new steering model.
Bugzilla: 737661 738491 739139 749059 755741 756147 756392
RH-Acked-by: John Feeney <jfeeney@redhat.com>

For Ethernet mode only,
When we want to register QP as promiscuous, it must be added to all the
existing steering entries and also to the default one.
The promiscuous QP might also be on of "real" QPs,
which means we need to monitor every entry to avoid duplicates and ensure
we close an entry when all it has is promiscuous QPs.
Same mechanism both for unicast and multicast.

Signed-off-by: Yevgeny Petrilin <yevgenyp@mellanox.co.il>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit b12d93d63c3217f0ec923ff938b12a744e242ffa)

Conflicts:

	drivers/net/mlx4/mlx4.h

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>