Revert "[infiniband] qib: add thresholds to VendorPortCounters PMA operation"
Patchwork-id: 45130
O-Subject: [Patch RHEL6 033/216] Revert "[infiniband] qib: add thresholds to
	VendorPortCounters PMA operation"
Bugzilla: 722308 738491 739139 749059 755741
RH-Acked-by: John Feeney <jfeeney@redhat.com>

This reverts commit 86b3d583457813a9824cd41d78054b0e1697683e.

This patch was submitted upstream over a year ago, however the original
submitter never drove it to final acceptance.  As a result, upstream
has moved on without this patch, the code that this patch modifies
has been usurped from the QLogic drivers and moved into a more general,
useful for any driver role, and the extension that this patch created
has not been carried to that new role.  So, drop this patch and pick
up the upstream changes.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>