IB/iser: Use separate buffers for the login request/response
Patchwork-id: 45191
O-Subject: [Patch RHEL6 125/216] IB/iser: Use separate buffers for the login
	request/response
Bugzilla: 738491 739139 749059 755741
RH-Acked-by: John Feeney <jfeeney@redhat.com>

From: Or Gerlitz <ogerlitz@mellanox.com>

The driver counted on the transactional nature of iSCSI login/text
flows and used the same buffer for both the request and the response.
We also went further and did DMA mapping only once, with
DMA_FROM_DEVICE, which violates the DMA mapping API.  Fix that by
using different buffers, one for requests and one for responses, and
use the correct DMA mapping direction for each.

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>
(cherry picked from commit 2c4ce609347f2a45792c8d9ebb5af11217766cb6)

Conflicts:

	drivers/infiniband/ulp/iser/iser_verbs.c

This patch differs from upstream considerably.  At the time this patch
was written, we still backed out various allocations in the error
path of the ib connection allocation routine.  Current upstream no
longer does that, they allow the tear down routine that's called on
the final put to do that for them.  So all the back out on error is
handled by the connection release routine.  Our kernel already did
this too, so keep with that.  However, upstream undoes the dma mapping
in what is the wrong place IMO, so this patch undoes the dma mapping
in the right place instead.

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>