IB/qib: Optimize locking for get_txreq()
Patchwork-id: 45265
O-Subject: [Patch RHEL6 184/216] IB/qib: Optimize locking for get_txreq()
Bugzilla: 722308 738491 739139 749059 755741
RH-Acked-by: John Feeney <jfeeney@redhat.com>

From: Mike Marciniszyn <mike.marciniszyn@qlogic.com>

The current code locks the QP s_lock, followed by the pending_lock, I
guess to to protect against the allocate failing.

This patch only locks the pending_lock, assuming that the empty case
is an exeception, in which case the pending_lock is dropped, and the
original code is executed.  This will save a lock of s_lock in the
normal case.

The observation is that the sdma descriptors will deplete at twice the
rate of txreq's, so this should be rare.

Signed-off-by: Mike Marciniszyn <mike.marciniszyn@qlogic.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>
(cherry picked from commit 489471095170ed1c6d0341739a243461638b0e06)

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>