blkfront: don't change to closing if we're busy
Patchwork-id: 45332
O-Subject: [RHEL6.3 PATCH 12/12] blkfront: don't change to closing if we're busy
Bugzilla: 773219
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=773219

We just reported to xenbus that we can't close yet, because we're still
in use. So we shouldn't then immediately tell xenbus that we are
closing. If we do, then we risk bypassing the chance to complain again,
if we're told to close again, and we're still in use. The reason this
code was here was to implement a deferred close, which doesn't work
on RHEL5 anyway (see bugs 794803 and 794806). It was added with upstream
commit b70f5fa043b (06/12 in this series). Besides being able to
complain each time, this patch fixes a problem when running on hosts
with legacy tools that may interpret the lack of a complaint or
state=Closing incorrectly, and then yank the device out from under the
guest.

I've sent a different version of this patch upstream that tries to
keep the deferred close feature and protect the guest. It hasn't been
accepted yet. For RHEL6, just making sure we restore the previous
behavior (pre patch 06/12) and not cause problems is the better route
anyway. So this is a RHEL-only patch now and the upstream result doesn't
matter.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>