GFS2: panic in gfs2_rindex_update during gfs2_grow
Patchwork-id: 45336
O-Subject: [RHEL6.3 PATCH] bz791010 - GFS2: panic in gfs2_rindex_update during
	gfs2_grow
Bugzilla: 791010
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch fixes bugzilla bug #791010. Description is given below.
This is a RHEL6 crosswrite from this upstream patch:

http://git.kernel.org/?p=linux/kernel/git/steve/gfs2-3.0-nmw.git;a=commitdiff;h=2b86c11111801aa12ada889ae90666774ec77100

The patch was tested on system gfs-a16c-02.mpc.lab.eng.bos.redhat.com
with a scenario that recreated faithfully without the patch.

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
--
GFS2: Ensure rindex is uptodate for fallocate master

This patch fixes a problem whereby gfs2_grow was failing and causing GFS2
to assert. The problem was that when GFS2's fallocate operation tried to
acquire an "allocation" it made sure the rindex was up to date, and if not,
it called gfs2_rindex_update. However, if the file being fallocated was
the rindex itself, it was already locked at that point. By calling
gfs2_rindex_update at an earlier point in time, we bring rindex up to date
and thereby avoid trying to lock it when the "allocation" is acquired.

rhbz#791010

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>