make bonding slaves honour master's skb->priority
Patchwork-id: 45344
O-Subject: [PATCH RHEL6.3] net: make bonding slaves honour master's skb->priority
Bugzilla: 749416
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

A user reported that traffic class 1 was being used when it should not
while using bonding and with FCoE.  This was because priority was being
overwritten when the slave was getting ready to send the frame.
Dropping the call to set the priority as upstream has done fixes this
properly.

This patch is a backport of the following upstream commit:

	commit 59fdaca9a4497ada47328d7b4b406b98a6f1c1a6
	Author: Maciej Żenczykowski <maze@google.com>
	Date:   Mon Oct 24 19:04:16 2011 +0000

	    net: make bonding slaves honour master's skb->priority

	    Signed-off-by: Maciej Żenczykowski <maze@google.com>
	    Acked-by: Flavio Leitner <fbl@redhat.com>
	    Signed-off-by: Jay Vosburgh <fubar@us.ibm.com>
	    Signed-off-by: David S. Miller <davem@davemloft.net>

This will resolve RHBZ 749416.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>