kabi: Add greylist check support for kernel rpm building
Patchwork-id: 45398
O-Subject: [PATCHv2 RHEL6.3 2/2] kabi: Add greylist check support for kernel
	rpm building
Bugzilla: 613726
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

description
===========
Adding support to monitor list of functions which are not
on the KABI whitelist, but partners use them inside their
modules and are interested in knowing wether they changed.
This is currently known as 'greylist'.

During the kernel rpm build, the whitelist data are checked,
and the build is aborted if the build mismatch the KABI.

Adding same framework and check for greylist, but letting
the build to continue in case of mismatch. The greylist
status is part of the build.log separated by following lines:

**** kABI: greylist checking start ****
...
**** kABI: greylist checking stop  ****

It'd be nice to have separate log file for KABI output, but
I'm not sure if that's more TODO or a wish.. :)

brew build
==========
https://brewweb.devel.redhat.com/taskinfo?taskID=4011655

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>