ext3: Return -EINVAL when start is beyond the end of fs in ext3_trim_fs()
Patchwork-id: 45421
O-Subject: [RHEL6.3 PATCH v2 06/10] ext3: Return -EINVAL when start is beyond
	the end of fs in ext3_trim_fs()
Bugzilla: 786987
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 786987
https://bugzilla.redhat.com/show_bug.cgi?id=786987

BREW 4013821
https://brewweb.devel.redhat.com/taskinfo?taskID=4013821

Simple backport of upstream commit:
2c2ea9451fc2a12ee57c8346f0da26969d07ee7f

We should return -EINVAL when the FITRIM parameters are not sane, but
currently we are exiting silently if start is beyond the end of the
file system. This commit fixes this so we return -EINVAL as other file
systems do.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
CC: Jan Kara <jack@suse.cz>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>