ext3: fix start and len arguments handling in ext3_trim_fs()
Patchwork-id: 45429
O-Subject: [RHEL6.3 PATCH v2 10/10] ext3: fix start and len arguments handling
	in ext3_trim_fs()
Bugzilla: 767219
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

BZ 767219
https://bugzilla.redhat.com/show_bug.cgi?id=767219

Upstream status:
The patch has been proposed upsrtream, but it have not been merged yet.

The overflow can happen when we are calling get_group_no_and_offset()
which stores the result of do_div() in 32 bit long type. However the
result might be bigger than that if big blocknr is passed in. This will
most likely happen when calling FITRIM with the default argument len =
ULLONG_MAX.

Fix this by using "end" variable instead of "start+len" as it is easier
to get right and specifically check that the end is not beyond the end
of the file system, so we are sure that the result of
get_group_no_and_offset() will not overflow. Otherwise truncate it to
the size of the file system.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>