kabi: remove mac80211 symbols
Patchwork-id: 45434
O-Subject: [RHEL 6 PATCH 001/102] kabi: remove mac80211 symbols
Bugzilla: 766952
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

It was a mistake to add these symbols in the first place.  The mac80211
bits are highly integrated with the individual drivers, and should be
treated as a driver library rather than as an API.

Without this, a rebase of the mac80211 subsystem is virtually
impossible.

bz766952

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>