ftrace: Use counters to enable functions to trace
Patchwork-id: 45535
O-Subject: Re: [RHEL6.3 BZ 454694 PATCH 17/32] [PATCH 17/32] ftrace: Use
	counters to enable functions to trace
Bugzilla: 454694
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Here's the updated patch, the rest should apply without changes.

>From c34a4f3b3d0a4d79e0194417b7e7ea29cdaa7955 Mon Sep 17 00:00:00 2001
From: Steven Rostedt <srostedt@redhat.com>
Date: Tue, 3 May 2011 13:25:24 -0400
Subject: [PATCH] ftrace: Use counters to enable functions to trace

Upstream: ed926f9b35cda0988234c356e16a7cb30f4e5338
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=454694

Every function has its own record that stores the instruction
pointer and flags for the function to be traced. There are only
two flags: enabled and free. The enabled flag states that tracing
for the function has been enabled (actively traced), and the free
flag states that the record no longer points to a function and can
be used by new functions (loaded modules).

These flags are now moved to the MSB of the flags (actually just
the top 32bits). The rest of the bits (30 bits) are now used as
a ref counter. Everytime a tracer register functions to trace,
those functions will have its counter incremented.

When tracing is enabled, to determine if a function should be traced,
the counter is examined, and if it is non-zero it is set to trace.

When a ftrace_ops is registered to trace functions, its hashes
are examined. If the ftrace_ops filter_hash count is zero, then
all functions are set to be traced, otherwise only the functions
in the hash are to be traced. The exception to this is if a function
is also in the ftrace_ops notrace_hash. Then that function's counter
is not incremented for this ftrace_ops.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>