fuse: add O_DIRECT support
Patchwork-id: 45543
O-Subject: [RHEL 6.3] fuse: add O_DIRECT support
Bugzilla: 753798
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

Hello,

This is a backport of the patch posted by Anand Avati.  It has been acked by
Miklos, the upstream FUSE maintainer, and is for bz 753798.  It is critical for
the gluster guys so reviews and acks appreciated, it's pretty straightforward
tho.  Thanks,

Josef

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>