cifs: fix dentry refcount leak when opening a FIFO on lookup
Patchwork-id: 45600
O-Subject: [RHEL6.3 PATCH BZ 781893] cifs: fix dentry refcount leak when
	opening a FIFO on lookup
Bugzilla: 781893
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Code analysis at
https://bugzilla.redhat.com/show_bug.cgi?id=781893#c5

Backport of the patch recently sent upstream by Jeff Layton to fix
a dentry refcount mismatch which resulted in a crash with the message
kernel BUG at fs/dcache.c:670!

http://thread.gmane.org/gmane.linux.kernel.cifs/5526

-
The cifs code will attempt to open files on lookup under certain
circumstances. What happens though if we find that the file we opened
was actually a FIFO or other special file?

Currently, the open filehandle just ends up being leaked leading to
a dentry refcount mismatch and oops on umount. Fix this by having the
code close the filehandle on the server if it turns out not to be a
regular file. While we're at it, change this spaghetti if statement
into a switch too.

Cc: stable@vger.kernel.org
Reported-by: CAI Qian <caiqian@redhat.com>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
-

This has been merged into Steve French's tree at
http://gitweb.samba.org/?p=sfrench/cifs-2.6.git;a=commit;h=5bccda0ebc7c0331b81ac47d39e4b920b198b2cd

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>