uv_mmrs.h cleanup patch
Patchwork-id: 45627
O-Subject: Re: [RHEL6.3 PATCH 16/15] uv_mmrs.h cleanup patch
Bugzilla: 737747
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

bz737747: uv_mmrs.h cleanup patch
First Post

Dean indicated that this would be much easier to review.
bz737747: uv_mmrs.h cleanup
First Post

The first patch in this series was partially applied in 6.2, but the
rest was held back because it was totally untested at that time but
has been now.

This following patch fragement was dropped both from the 6.2 patch
and the patch first in this series.  The upstream patch is

    commit 55ba41202866f4e135c39d0bb88f128b96ba2167
    Author: Jack Steiner <steiner@sgi.com>
    Date:   Fri May 27 09:52:56 2011 -0500

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>