sd_dif: fix setting bio flags
Patchwork-id: 45676
O-Subject: [RHEL6 PATCH] sd_dif: fix setting bio flags
Bugzilla: 799075
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

Hi,

A patch was just posted upstream to fix the setting of flags for sd_dif
and floppy.c:

http://groups.google.com/group/linux.kernel/browse_thread/thread/87487f9987731480/da7877b036bfac34?show_docid=da7877b036bfac34&pli=1

The problem is a simple thinko, setting flags to BIO_* instead of
(1<<BIO_*).  I did a grep through our tree to see if there were any
other problems of this type and didn't find any.  This issue was
introduced to our kernel with this commit (which I acked... I'm so
ashamed):

commit 57b450ad7c971adfec985ac4018a3dee2a1dbb71
Author: Mike Snitzer <snitzer@redhat.com>
Date:   Thu Apr 21 21:04:31 2011 -0400

    [block] Make the integrity mapped property a bio flag

That commit went out in 6.2, so, depending on how important we think
this is, it may need a 6.2.z fix as well (opinions on this welcome).

Upstream, Linus said he'd take this in soon, and Jens and Martin seemed
ok with the change.

This addresses bug 799075
(https://bugzilla.redhat.com/show_bug.cgi?id=799075).

Cheers,
Jeff

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>