NFSv4: Handle expired stateids when the lease is still valid
Patchwork-id: 45707
O-Subject: [PATCH RHEL 6.3] NFSv4: Handle expired stateids when the lease is
	still valid
Bugzilla: 757876
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Fixes BZ 757876

Netapp reported an issue on NFSv4.1 testing.
The problem described in their words
--
If the server returns NFS4ERR_EXPIRED in reply to a READ or WRITE, the
state_manager fires off a solo SEQUENCE operation to test the lease with
out handling the expired stateid. The SEQUENCE operation succeeds, so
the lease is assumed to be ok and no lease recovery occurs. We fail to
recover from the expired stateid and end up looping forever in a
READ/WRITE + solo SEQUENCE death spiral.
--

This problem was initially reported upstream for a NFSv4 error where the
server returns a NFS4ERR_EXPIRED error. A subsequent RENEW operation
returned NFS4_OK which means the state is never renewed. This results
in the same stateid being used and the READ-EXPIRED-RENEW cycle being
repeated.
http://thread.gmane.org/gmane.linux.nfs/40460

The patch works by forcing a OPEN to obtain a new state when a
NFS4ERR_EXPIRED is returned by the server.

Backport of upstream patch
--
commit 0ced63d1a245ac11241a5d37932e6d04d9c8040d
Author: Trond Myklebust <Trond.Myklebust@netapp.com>
Date:   Thu May 26 14:26:35 2011 -0400
[PATCH] NFSv4: Handle expired stateids when the lease is still valid

Currently, if the server returns NFS4ERR_EXPIRED in reply to a READ or
WRITE, but the RENEW test determines that the lease is still active, we
fail to recover and end up looping forever in a READ/WRITE + RENEW death
spiral.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: stable@kernel.org
--

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>