GFS2: rename existing file after mount crashes
Patchwork-id: 45710
O-Subject: [RHEL6.3 PATCH] bz 794839 - GFS2: rename existing file after mount
	crashes
Bugzilla: 794839
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch fixes this bug:
794839 - GFS2: rename existing file after mount crashes
This is a RHEL6 crosswrite that combines the following two
upstream patches:
http://git.kernel.org/?p=linux/kernel/git/steve/gfs2-3.0-nmw.git;a=commitdiff;h=58884c4df005ee5ee854cfcd0385d5a6bf25aa30
http://git.kernel.org/?p=linux/kernel/git/steve/gfs2-3.0-nmw.git;a=commitdiff;h=a365fbf354907430e6852f0c373b4b3eeff81ba3

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
--
GFS2: rename existing file after mount crashes

GFS2 is crashing on renames immediately after mount.  This is because it isn't
reading in the rindex file.  This patch makes GFS2 always read in the rindex
file on mount.  When it reads in the rindex file, it also checks if it is
currently holding the rindex lock, to avoid grabbing it recursively. It
also adds code to check if the rgrp information is current, and if not
(e.g. due to gfs2_grow) it re-reads it.

rhbz#794839

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>