fcoe: Only define ndo_fcoe_get_hbainfo if fcoe is configured
Patchwork-id: 45721
O-Subject: [RHEL 6.3 PATCH] fcoe: Only define ndo_fcoe_get_hbainfo if fcoe is
	configured
Bugzilla: 789086
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

From: Neil Horman <nhorman@tuxdriver.com>

s390 is getting a bunch of warnings because  netdev_fcoe_hbainfo is used in a
parameter list but not defined (due to CONFIG_FCOE not being defined).  While it
probably should be, thats an upstream debate.  Upstream protects the calling
function pointer (ndo_fcoe_get_hbainfo) by wrapping it in CONFIG_FCOE.  While
I'm not thrilled with doing this (since we have to define it in the
net_device_extended area),  It won't hurt anything in this case, and will
silence the warnings.

This additional patch to the series resolves the warnings that Aris asked me to
correct.

Resolves: bz 789086

Brew:
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=4120697

Testing:
Did a brew build above on s390 to ensure the warnings were gone, and a local
x86_64 build to ensure nothing else broke

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>