iser: post initial receive buffers before sending the final login request
Patchwork-id: 45727
O-Subject: [RHEL6.3 PATCH 1/1] IB/iser: post initial receive buffers before
	sending the final login request
Bugzilla: 800041
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 800041.

>From the upstream commit
http://git.kernel.org/?p=linux/kernel/git/roland/infiniband.git;a=commitdiff;h=89e984e2c2cd14f77ccb26c47726ac7f13b70ae8:

An iser target may send iscsi NO-OP PDUs as soon as it marks the iSER
iSCSI session as fully operative.  This means that there is window
where there are no posted receive buffers on the initiator side, so
it's possible for the iSER RC connection to break because of RNR NAK /
retry errors.  To fix this, rely on the flags bits in the login
request to have FFP (0x3) in the lower nibble as a marker for the
final login request, and post an initial chunk of receive buffers
before sending that login request instead of after getting the login
response.

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Roland Dreier <roland@purestorage.com>

I have tested the patch by doing login/logout tests with
iser. I was not able to hit the same problem described.
I have relied on Mellanox for that testing.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>