add RHELPatchRules, guidelines for RHKL patch postings
Patchwork-id: 4724
O-Subject: [kernel team] [RHEL6 INTERNAL PATCH] add RHELPatchRules,
	guidelines for RHKL patch postings
Bugzilla:
RH-Acked-by: Mauro Carvalho Chehab <mchehab@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Thomas Graf <tgraf@redhat.com>

Red Hat Kernel Engineering has grown significantly over the years, and
the number of patches being posted on RHKL has grown proportionally.  It
has become impossible to manually track the large number of patches
posted to RHKL, so the kernel maintainers developed a sophisticated
toolset that allows for full auditing of how customer issues were
resolved, and how patches from those customer issues got into the RHEL
kernel trees.

Like any utility, the toolset developed by the maintainers requires a
format that developers should follow in order to help both their fellow
engineers review patches, and help the maintainers process patches as
quickly and as efficiently as possible.  Obviously, both of these groups
must be satisified in order for patches to be committed to the RHEL
kernel trees.

The document is an attempt to gather the toolset's _current_ required
formatting of patches (beyond that which upstream's
Documentation/SubmittingPatches requires) into a set of
guidelines that are useful for patch authors to follow when submitting
patches to RHKL.  This document does NOT introduce new requirements upon
the engineering staff; all of the guidelines have been enforced
previously on RHKL.

This document was written in collaboration with the current RHEL
maintainers and received input from engineers in Westford, Raleigh/RDU,
Brno, remote engineers, as well as a few onsite partner engineers.

Neil Horman is currently working on a utility similar to upstream's
checkpatch.pl script which will verify a patch before it is posted to
RHKL.

I'd like to point out the efforts of Aris Rozanski, Jarod Wilson, Kyle
McMartin, and Don Zickus in writing this document.  It would have been
impossible to write this document without their input and guidance.

Thanks to everyone who pitched in a review and a big thanks to those
that added a Reviewed-by line.

P.