KEYS: Do preallocation for __key_link()
Patchwork-id: 45783
O-Subject: [RHEL6 PATCH 05/38] KEYS: Do preallocation for __key_link()
Bugzilla: 788634
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Do preallocation for __key_link() so that the various callers in request_key.c
can deal with any errors from this source before attempting to construct a key.
This allows them to assume that the actual linkage step is guaranteed to be
successful.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: James Morris <jmorris@namei.org>